rename + rehaul
index.js -> gcindex.js 
file was partially rewritten and reformatted to change the buttons to checkboxes. In documentation they will still be referred to as buttons, but they behave as checkboxes and will be treated as checkboxes in the code. 
Also added behavior to limit user input to 5 selected genres and 5 selected movies.